add support for double

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 
 int fib(int n)
 {
-	long long res;
+	double res;
 
 	if ((n == 1) || (n == 2))
 	{
